requirements.txt
imutils==0.5.4
numba==0.55.1
numpy==1.21.6
opencv_python==4.5.5.64
PyYAML==6.0
yacs==0.1.8
